DBaaS/VPC Integration: Implementation (#714)
* WIP

* done

* VPC & VPC Subnet deletion retries

* Refactor should_retry_subnet_delete_400s

* Fix tests

* Bump linode-api4 to v5.37.0

* Update plugins/module_utils/linode_helper.py


---------

+++ plugins/modules/vpc_subnet.py
@@ -18,7 +18,11 @@ from ansible_collections.linode.cloud.plugins.module_utils.linode_docs import (
 from ansible_collections.linode.cloud.plugins.module_utils.linode_helper import (
     filter_null_values,
     handle_updates,
+    retry_on_response_status,
     safe_find,
+)
+from ansible_collections.linode.cloud.plugins.module_utils.linode_vpc_shared import (
+    should_retry_subnet_delete_400s,
 )
 from ansible_specdoc.objects import (
     FieldType,
@@ -135,7 +139,15 @@ class Module(LinodeModuleBase):
         )
         if subnet is not None:
             self.results["subnet"] = subnet._raw_json
-            subnet.delete()
+
+            # If any entities attached to this subnet are in a transient state
+            # expected to eventually allow deletions,
+            # retry the delete until it succeeds.
+            if should_retry_subnet_delete_400s(self.client, subnet):
+                retry_on_response_status(self._timeout_ctx, subnet.delete, 400)
+            else:
+                subnet.delete()
+
             self.register_action(f"Deleted VPC Subnet {label}")
 
     def exec_module(self, **kwargs: Any) -> Optional[dict]:
